requirements.txt
streamlit>=1.18.0
pandas>=2.0.0
numpy>=1.23.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
tensorflow>=2.14.0
Pillow>=9.0.0
wordcloud>=1.9.0
openai>=0.28
yfinance>=0.2.0
python-docx>=0.8.11
pdfplumber>=0.9.0
joblib>=1.2.0
